Fix clang-tidy warnings in test code
- Cast uint16_t CRC values to unsigned before bitwise shifts to
  avoid hicpp-signed-bitwise (uint16_t promotes to signed int)
- Initialize local variable to satisfy cppcoreguidelines-init-variables
- Add parentheses around MEGA multiplications for readability-math-missing-parentheses
- Split multiple declarations to satisfy readability-isolate-declaration
- Use unsigned loop variable to avoid bugprone-too-small-loop-variable

+++ tests/src/test_intrusive_util.c
@@ -633,8 +633,8 @@ static void test_crc_residue_property(void)
     // Append CRC in big-endian order.
     uint8_t augmented[sizeof(data) + CRC_SIZE_BYTES];
     memcpy(augmented, data, sizeof(data));
-    augmented[sizeof(data)]     = (uint8_t)(crc >> 8U);   // high byte
-    augmented[sizeof(data) + 1] = (uint8_t)(crc & 0xFFU); // low byte
+    augmented[sizeof(data)]     = (uint8_t)((unsigned)crc >> 8U);   // high byte
+    augmented[sizeof(data) + 1] = (uint8_t)((unsigned)crc & 0xFFU); // low byte
     const uint16_t residue      = crc_add(CRC_INITIAL, sizeof(augmented), augmented);
     TEST_ASSERT_EQUAL_HEX16(CRC_RESIDUE, residue);
